backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.43

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=43.0.0

# HTTP Clients & Web Requests
httpx>=0.26.0
certifi>=2024.0.0
aiohttp>=3.9.3
requests>=2.31.0

# AI & Analysis
openai>=1.10.0
tiktoken>=0.5.2

# Utilities & Logging
loguru==0.7.2
tenacity==8.2.3

# Caching & Background Tasks
redis>=5.0.1
celery==5.3.6

# Web Scraping & Content Extraction
beautifulsoup4==4.12.3
trafilatura==1.8.0
playwright==1.41.1
scrapy==2.11.0
lxml>=4.9.4
html5lib==1.1
selectolax==0.3.21

# File Processing & Storage
python-magic==0.4.27
Pillow>=10.2.0
aiofiles>=23.2.1
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.2

# Google APIs & Services
google-api-python-client>=2.179.0
google-auth>=2.40.3
google-auth-oauthlib>=1.2.2
google-auth-httplib2>=0.2.0

# Google Ads API (Critical for keyword metrics)
google-ads>=25.1.0
grpcio>=1.62.0,<1.74.0

# Google Cloud (for BigQuery integration)
google-cloud-bigquery>=3.14.1
google-cloud-pubsub>=2.19.4

# Video & Media Processing
isodate>=0.7.2

# Scheduling & Cron
croniter>=2.0.1
cron-descriptor>=1.4.0

# Data Processing & Validation
email-validator>=2.3.0
validators>=0.22.0

# Advanced Data Processing
pandas>=2.2.0
numpy>=1.26.3

# Rate Limiting
aiolimiter>=1.1.0
slowapi>=0.1.9

# Caching
cachetools>=5.5.2

# Structured Logging
structlog>=24.1.0

# Generic Dimensions Feature Dependencies
textstat>=0.7.3
spacy>=3.4.0
jsonschema>=4.20.0

# Additional Dependencies (discovered during deployment)
dnspython>=2.7.0
googleapis-common-protos>=1.70.0
httplib2>=0.30.0
oauthlib>=3.3.1
proto-plus>=1.26.1
protobuf>=6.32.0
pyparsing>=3.2.3
pyasn1-modules>=0.4.2
requests-oauthlib>=2.0.0
uritemplate>=4.2.0

# Async Redis Support
redis[hiredis]>=5.0.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpretty>=1.1.4

# Development
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0

# Additional utilities
python-dateutil>=2.8.2
urllib3>=2.1.0
charset-normalizer>=3.3.2
idna>=3.6